requirements.txt
--index-url https://pypi.fury.io/zerthmonk/
--extra-index-url https://pypi.org/simple/

wheel>=0.35.1
paho-mqtt>=1.4.0
pyaml>=18.0.0
filelock>=3.0.12,<3.1
pygame==2.6.0
pytest>=6.1.2
requests==2.32.3
flake8>=3.8.4
autoflake
isort
